+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.0-11 / Unreleased
-  *
+  * [Added] /dist/jquery.fancytree-custom.min.js with AMD support
+  
 
 # 2.0.0-10 / 2014-04-13
   * [Added] New method node.appendSibling()