+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   * [Added] /dist/jquery.fancytree-custom.min.js with AMD support
   * [Added] #56: Allow to set special node.attributes from data-...
   * [Added] #191: Allow to set additional tree.data attributes from <ul data-...>
+  * [Improved] #192: Removed 'height: 100%' for container (was introduced to fix 
+     an IE 9 bug, that now should be solved by 'min-height: 0%')
   * [Fixed] #197: Allow special characters in tooltips