+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# 2.0.0-11 / Unreleased
+# 2.0.0-12 / Unreleased
+  * [Added] /dist/src folder contains uncompressed extensions for bower
+
+
+# 2.0.0-11 / 2014-04-27
   * [Added] /dist/jquery.fancytree-custom.min.js with AMD support
   * [Added] #56: Allow to set special node.attributes from data-...
   * [Added] #191: Allow to set additional tree.data attributes from <ul data-...>